Updated pdbqt_split.sh
There was a bug in the original script that would overlook a great deal of pdbqt files, which led to a reduction of sample files. This was because the original script called on the autodock program to split up the files. The new script no longer relies on autodock and uses simple logic to split the input files into the necessary chunks for the virtual screening.